Move tests to root, simplifies ava config
