+++ test/usageNotes-test.js
@@ -2,7 +2,7 @@
 
 var test = require('ava');
 var tempWrite = require('temp-write');
-var usageNotes = require('../lib/usageNotes.js');
+var usageNotes = require('../server/lib/usageNotes.js');
 
 test('usageNotes', function (t) {
     var filepath = tempWrite.sync('this is some test text');
